agents/lecture_insight/requirements.txt
# Lecture Insight Agent - Requirements
# Only packages actually used by this agent

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0

# Environment & Config
python-dotenv>=1.0.0

# Database
aiosqlite>=0.19.0

# AI/ML Workflow
langgraph>=0.2.0
google-generativeai>=0.3.0

# Audio Transcription
assemblyai>=0.17.0

# Search APIs
tavily-python>=0.3.0
google-api-python-client>=2.0.0
